requirements.txt
Flask==2.3.3
tensorflow==2.13.0
gunicorn==21.2.0
pandas==2.0.3
numpy==1.24.3
Pillow==10.0.0
Werkzeug==2.3.7
flask-cors==4.0.0
google-generativeai==0.3.1